backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
reportlab==4.2.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.7.0
PyPDF2==3.0.1
python-docx==1.1.0
sendgrid==6.11.0
